Tidy refactor
